Changes to card issuances screen

+++ oauth/src/main/java/jp/co/soramitsu/oauth/base/sdk/contract/SoraCardContractData.kt
@@ -16,5 +16,5 @@ data class SoraCardContractData(
     val environment: SoraCardEnvironmentType,
     val soraCardInfo: SoraCardInfo?,
     val client: String,
-    val userAvailableXorAmount: Float,
+    val userAvailableXorAmount: Double,
 ) : Parcelable
